+++ layers/GatoGraphQLForWP/plugins/gatographql/src/ModuleResolvers/Extensions/PremiumExtensionModuleResolver.php
@@ -12,6 +12,7 @@ class PremiumExtensionModuleResolver extends AbstractExtensionModuleResolver
 
     public const AUTOMATION = Plugin::NAMESPACE . '\\extensions\\automation';
     public const CHATGPT_TRANSLATION = Plugin::NAMESPACE . '\\extensions\\chatgpt-translation';
+    public const CLAUDE_TRANSLATION = Plugin::NAMESPACE . '\\extensions\\claude-translation';
     public const DEEPL = Plugin::NAMESPACE . '\\extensions\\deepl';
     public const EVENTS_MANAGER = Plugin::NAMESPACE . '\\extensions\\events-manager';
     public const GOOGLE_TRANSLATE = Plugin::NAMESPACE . '\\extensions\\google-translate';
@@ -27,6 +28,7 @@ class PremiumExtensionModuleResolver extends AbstractExtensionModuleResolver
         return [
             self::AUTOMATION,
             self::CHATGPT_TRANSLATION,
+            self::CLAUDE_TRANSLATION,
             self::DEEPL,
             self::EVENTS_MANAGER,
             self::GOOGLE_TRANSLATE,
@@ -41,6 +43,7 @@ class PremiumExtensionModuleResolver extends AbstractExtensionModuleResolver
         return match ($module) {
             self::AUTOMATION => \__('Automation', 'gatographql'),
             self::CHATGPT_TRANSLATION => \__('ChatGPT Translation', 'gatographql'),
+            self::CLAUDE_TRANSLATION => \__('Claude Translation', 'gatographql'),
             self::DEEPL => \__('DeepL', 'gatographql'),
             self::EVENTS_MANAGER => \__('Events Manager', 'gatographql'),
             self::GOOGLE_TRANSLATE => \__('Google Translate', 'gatographql'),
@@ -56,6 +59,7 @@ class PremiumExtensionModuleResolver extends AbstractExtensionModuleResolver
         return match ($module) {
             self::AUTOMATION => \__('Use GraphQL to automate tasks in your app: Execute queries when some event happens, chain queries, and schedule and trigger queries via WP-Cron. (The Internal GraphQL Server extension is required).', 'gatographql'),
             self::CHATGPT_TRANSLATION => \__('Translate content to multiple languages using the ChatGPT API.', 'gatographql'),
+            self::CLAUDE_TRANSLATION => \__('Translate content to multiple languages using the Claude API.', 'gatographql'),
             self::DEEPL => \__('Translate content to multiple languages using the DeepL API.', 'gatographql'),
             self::EVENTS_MANAGER => \__('Integration with plugin "Events Manager", adding fields to the schema to fetch event data.', 'gatographql'),
             self::GOOGLE_TRANSLATE => \__('Translate content to multiple languages using the Google Translate API.', 'gatographql'),